dasdas

+++ project_back/services/task_service.js
@@ -1,7 +1,9 @@
 import taskRepo from "../repositories/task_repository.js";
 import projectService from "./project_service.js";
- 
-
+import Attachment from "../models/Attachment.js";
+import TagTask from "../models/TaskTag.js";
+import Comment from "../models/Comment.js";
+import TaskAssignment from "../models/TaksAssignment.js";
 class TaskService {
     async createTask(data) {
         return await taskRepo.createTask(data);
@@ -34,7 +36,22 @@ class TaskService {
     }
 
     async deleteTask(projectId, taskId) {
-        return await taskRepo.deleteTask(projectId, taskId);
+        
+        const task = await taskRepo.findTaskById(taskId);
+        if (!task || task.projectId.toString() !== projectId) {
+            throw new Error('Tarea no encontrada o no pertenece al proyecto');
+        }
+
+        
+        await taskRepo.deleteTask(projectId, taskId);
+
+        
+        await Promise.all([
+            Comment.deleteMany({ taskId }),
+            TaskAssignment.deleteMany({ taskId }),
+            TagTask.deleteMany({ taskId }),
+            Attachment.deleteMany({ taskId }),
+        ]);
     }
     async markTaskAsCompleted(taskId) {
         return await taskRepo.markTaskAsCompleted(taskId);
